CD/scripts/requirements.txt
selenium
python-owasp-zap-v2.4
requests
webdriver-manager